servidor/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
grpcio
protobuf
grpcio-tools
googleapis-common-protos
numpy
pandas
matplotlib
seaborn
scikit-learn